refactor: move update builder to its own package
